api/DebatIDOAPI/requirements.txt
connexion >= 2.6.0; python_version>="3.6"
connexion >= 2.3.0; python_version=="3.5"
connexion >= 2.3.0; python_version=="3.4"
connexion == 2.4.0; python_version<="2.7"
swagger-ui-bundle >= 0.0.2
python_dateutil >= 2.6.0
setuptools >= 21.0.0

mysql-connector-python
pytz
sqlalchemy
flask-sqlalchemy
flask-cors
mysqlclient
